+++ src/main/java/io/chocorean/authmod/authentication/datasource/DatabaseSourceStrategy.java
@@ -1,4 +1,4 @@
-package io.chocorean.authmod.authentication;
+package io.chocorean.authmod.authentication.datasource;
 
 import io.chocorean.authmod.AuthMod;
 import io.chocorean.authmod.authentication.db.ConnectionFactory;
@@ -17,7 +17,7 @@ import java.sql.SQLException;
 public class DatabaseSourceStrategy implements IDataSourceStrategy {
 
     private final IPlayersDAO playersDAO;
-    public static final Logger LOGGER = FMLLog.getLogger();
+    public static final Logger LOGGER = FMLLog.log;
 
     public DatabaseSourceStrategy(AuthModDatabaseConfig config) {
         this.playersDAO = new PlayersDAO(ConnectionFactory.getConnection(), config.getTable());
