+++ src/main/java/io/chocorean/authmod/command/RegisterCommand.java
@@ -1,8 +1,7 @@
 package io.chocorean.authmod.command;
 
 import io.chocorean.authmod.authentication.AuthModule;
-import io.chocorean.authmod.authentication.IAuthenticationStrategy;
-import io.chocorean.authmod.authentication.IDataSourceStrategy;
+import io.chocorean.authmod.authentication.datasource.IDataSourceStrategy;
 import io.chocorean.authmod.event.Handler;
 import io.chocorean.authmod.model.IPlayer;
 import io.chocorean.authmod.model.Player;
